Authorization with Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -65,7 +65,10 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
-    {
+    {   
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied. You can't edit this question") ;
+        }
         return view("questions.edit", compact( 'question' )) ;
     }
 
@@ -78,6 +81,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied. You can't update this question") ;
+        }
+
         $question->update( $request->only('title', 'body') ) ;
 
         return redirect('/questions')->with( 'success', "Your question has been updated." ) ;
@@ -91,6 +99,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied. You can't delete this question") ;
+        }
+
         $question->delete() ;
 
         return redirect('/questions')->with( 'success', "Your question has been deleted." ) ;
